refactor(pack): refactor xdsl reset
