+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/reset/index.js
@@ -1,0 +1,32 @@
+import angular from 'angular';
+import ngOvhTelecomUniverseComponents from '@ovh-ux/ng-ovh-telecom-universe-components';
+import '@uirouter/angularjs';
+import '@ovh-ux/ng-translate-async-loader';
+import 'angular-translate';
+import 'ovh-api-services';
+
+import controller from './reset.controller';
+import template from './reset.html';
+
+const moduleName = 'XdslModemReset';
+
+angular
+  .module(moduleName, [
+    ngOvhTelecomUniverseComponents,
+    'ngTranslateAsyncLoader',
+    'pascalprecht.translate',
+    'ovh-api-services',
+    'ui.router',
+  ])
+  .controller('XdslModemResetCtrl', controller)
+  .run(/* @ngTranslationsInject:json ./translations */)
+  .run(
+    /* @ngInject */ ($templateCache) => {
+      $templateCache.put(
+        'app/telecom/pack/xdsl/modem/reset/pack-xdsl-modem-reset.html',
+        template,
+      );
+    },
+  );
+
+export default moduleName;
