+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/reset/reset.html
